In-progress

+++ check/src/board_builder.hpp
@@ -43,8 +43,8 @@ namespace wisdom
         std::vector<BBPieceWithCoordState> pieces_with_coords;
         std::vector<BBEnPassantState> en_passant_states;
         std::vector<BBCastlingState> castle_states;
-        int16_t half_moves_clock = 0;
-        int16_t full_moves = 0;
+        int half_moves_clock = 0;
+        int full_moves = 0;
 
     public:
         BoardBuilder () = default;
@@ -67,9 +67,9 @@ namespace wisdom
 
         void set_castling (Color who, CastlingState state);
 
-        void set_half_moves (int16_t new_half_moves_clock);
+        void set_half_moves (int new_half_moves_clock);
 
-        void set_full_moves (int16_t new_full_moves);
+        void set_full_moves (int new_full_moves);
 
         Board build ();
     };
